jwt authentication setup

+++ src/main/java/com/abhiraj/SponsorFlow/domain/dtos/request/BrandRequestDto.java
@@ -9,6 +9,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @AllArgsConstructor
 @NoArgsConstructor
 @Data
@@ -24,6 +26,6 @@ public class BrandRequestDto {
     @NotNull
     @Positive
     @Min(0)
-    private Double totalBudget;
+    private BigDecimal totalBudget;
 
 }
