+++ src/main/java/com/abhiraj/SponsorFlow/domain/dtos/response/BrandResponseDto.java
@@ -5,6 +5,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -13,6 +15,6 @@ public class BrandResponseDto {
 
     private Long id;
     private String name;
-    private Double availableBudget;
+    private BigDecimal availableBudget;
 
 }
