+++ src/main/java/com/abhiraj/SponsorFlow/domain/dtos/response/InfluencerResponseDto.java
@@ -5,6 +5,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -16,6 +18,6 @@ public class InfluencerResponseDto {
     private String platform;
     private Long followerCount;
     private Double engagementRate;
-    private Double totalEarnings;
+    private BigDecimal totalEarnings;
 
 }
